requirements.txt
adal==1.2.2
amqp==2.5.1
asn1crypto==0.24.0
azure-common==1.1.23
azure-keyvault==1.1.0
azure-nspkg==3.0.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
importlib-metadata==0.23
isodate==0.6.0
kombu==4.6.4
more-itertools==7.2.0
msrest==0.6.10
msrestazure==0.6.2
oauthlib==3.1.0
pycparser==2.19
PyJWT==1.7.1
pyodbc==4.0.27
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
